src/pyproject.toml
[project]
name = "solar-sentinel"
version = "0.1.0"
description = "Autonomous solar panel defect detection and classification using computer vision and NLP"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "aiosqlite>=0.20.0",
    "apscheduler>=3.10.0",
    "ultralytics>=8.3.0",
    "crewai>=1.0.0",
    "google-genai>=1.0.0",
    "aiosmtplib>=3.0.0",
    "python-telegram-bot>=21.0",
    "python-multipart>=0.0.18",
    "pillow>=11.0.0",
    "psutil>=6.0.0",
    "httpx>=0.28.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
# NOTE: picamera2 is NOT listed here because its transitive dep (python-prctl)
# is Linux-only and breaks cross-platform lockfile resolution.
# On the Pi, install it separately: uv pip install picamera2
# The app handles its absence gracefully (camera.py uses a stub on non-Pi).
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "httpx>=0.28.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
